+++ packages/examples/examples/src/components/basics/HiddenDiv.vue
@@ -22,11 +22,12 @@
               center,
               zoom,
             },
+            theme,
           }"
           :width="width"
           :height="height"
         >
-          <yandex-map-default-scheme-layer :settings="{ theme }" />
+          <yandex-map-default-scheme-layer />
           <yandex-map-controls :settings="{ position: 'right' }">
             <yandex-map-zoom-control />
           </yandex-map-controls>