+++ packages/examples/examples/src/components/basics/MapCamera.vue
@@ -14,11 +14,12 @@
             center,
             zoom,
           },
+          theme,
         }"
         :width="width"
         :height="height"
       >
-        <yandex-map-default-scheme-layer :settings="{ theme }" />
+        <yandex-map-default-scheme-layer />
         <yandex-map-controls :settings="{ position: 'right' }">
           <yandex-map-geolocation-control />
         </yandex-map-controls>