+++ packages/examples/examples/src/components/objects/HideMarkers.vue
@@ -13,11 +13,12 @@
             zoom,
           },
           zoomRange: ZOOM_RANGE,
+          theme,
         }"
         :width="width"
         :height="height"
       >
-        <yandex-map-default-scheme-layer :settings="{ theme }" />
+        <yandex-map-default-scheme-layer />
         <yandex-map-default-features-layer />
         <yandex-map-controls :settings="{ position: 'right' }">
           <yandex-map-zoom-control />